// src/app/page/master/matauang/matauang.interface.ts
import {EventEmitter} from '@angular/core';

export interface Matauang {
	kode?: any;
	id?: any;
	mataUang?: any;
	kdProfile?: string;
	namaModulAplikasi?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}